requirements.txt
jieba
librosa==0.10.2
einops
tqdm
fast_langdetect>=0.3.1
split-lang
transformers
cn2an
pypinyin
jieba_fast
wordsegment
onnxruntime-gpu
torchcodec
pyopenjtalk
g2p_en
ko_pron
g2pk2
eunjeon
ToJyutping